requirements.txt
anyio==4.6.2.post1
asgiref==3.8.1
attrs==24.2.0
beautifulsoup4==4.12.3
bs4==0.0.2
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
click==8.1.7
colorama==0.4.6
colourmap==1.1.17
contourpy==1.3.0
crontab==1.0.1
cycler==0.12.1
Django==5.1.3
django-cors-headers==4.6.0
django-cron==0.6.0
django-rest-framework==0.1.0
django-widget-tweaks==1.5.0
djangorestframework==3.15.2
dnspython==2.7.0
email_validator==2.2.0
fonttools==4.54.1
google-trends-api==2.2.6
googlesearch-python==1.2.5
googletrends==0.1.2
gtrends==0.2.2
h11==0.14.0
httpcore==1.0.7
httpx==0.27.2
idna==3.10
joblib==1.4.2
kiwisolver==1.4.7
loguru==0.7.2
lxml==5.3.0
matplotlib==3.9.2
mysqlclient==2.2.5
nltk==3.9.1
numpy==2.1.3
outcome==1.3.0.post0
packaging==24.1
pandas==2.2.3
pillow==11.0.0
pycparser==2.22
PyMySQL==1.1.1
pyparsing==3.2.0
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytrends==4.9.2
pytz==2024.2
rake-nltk==1.0.6
RapidFuzz==3.10.1
regex==2024.11.6
requests==2.32.3
scikit-learn==1.5.2
scipy==1.14.1
seaborn==0.13.2
selenium==4.26.1
setuptools==75.3.0
six==1.16.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.6
sqlparse==0.5.1
svgpathtools==1.6.1
svgwrite==1.4.3
tabulate==0.9.0
tenacity==9.0.0
threadpoolctl==3.5.0
tiki==0.2.0
tqdm==4.67.0
trendpy==1.0.0
trio==0.27.0
trio-websocket==0.11.1
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
webdriver-manager==4.0.2
websocket-client==1.8.0
wget==3.2
win32-setctime==1.1.0
worldmap==0.1.7
wsproto==1.2.0



